Editor::StartFileSelection now takes a std::function to allow for lambda use and to get rid of the crusty function that StartFileSelection used to be.
The FileSelect window will now keep track of the last directory that
was opened when previously selecting a file to make selecting multiple
files in the same location faster.

Editor::OpenPopup now takes a string for the popup's name and stores it
in a string so it has its own copy.

These functions have also received cleanup and should be easier to read.