The AssetLibrary has been overhauled. Before, the AssetLibrary had duplicate code to deal with different Asset types. This has been removed using templates.
Every Asset type contains resources and resources are expected to have
certain members. Gfx::Model and Gfx::Shader are currently the only
examples of resources. The members that make them usable as Asset
resources are located at the top of both of their definitions. In
general, the members provide the ability to initialize, purge, and check
whether the resource is live.

This is the foundation of what should eventually become Asset streaming.
Since we can track whether an Asset is live or not, we should be able to
make an initialization request and have that initialization happen on a
different thread as to not block the main thread.