The AssetLibrary now has functions for loading and saving assets.
Assets will be loaded when Varkor first launches and saved when Varkor
is closed. The AssetLibrary will also initialize the loaded assets so
they can be used instantly. This will be replaced with dynamically
initialized assets later to prevent slow startup times for projects with
a lot of assets.

The Vlk::Explorer also takes in a Pair by reference so it is more clear
that the Pair is not modified.