A file named components.vlk has been added to keep track of the components that are expected to be registered when Varkor runs.
When Varkor first launches, the registered types will be compared with
what is contained inside of components.vlk and any differences that are
found will be logged as errors.

This also includes a way to update the assets.vlk file and the
components.vlk file from the editor. Additionally, the Asset window is
opened by default when Varkor is launched.