Entire Spaces can now be serialized and deserialized. All of the previous commits pertaining to serialization and deserialization have been leading up to this one.
Under the File menu, there are now options for loading and saving a
spaces. Theses options will open up a file browser to perform the
action. Loading a space will just open up a browser that allows a user
to select a file. Saving a space will do the same but also allow the
user to change the name of the file that is being saved in a text box.

This also introduces VSerialize and VDeserialize component type
functions. Components must define these functions if they support the
saving and loading of data. Model and Transform are the only components
currently in use and they make use of these functions.

This also makes one minor change to space and member names. The default
names no longer have a space before the id. This makes things consistent
with asset names. It also means that filenames will not contain spaces
if the default space name is used. Spaces in filenames gets a hard no
from me.