Ds::RbTree now has a function for swapping nodes.
No part of the implementation makes use of the SwapNodes function, but
it has been tested within the RemoveNode function. Instead of moving a
value from a node, the nodes were swapped and all tests were still
passing.

This was introduced because I came across a case where swapping the
nodes seemed more appropriate. When a node is removed from the tree,
we replace the value within that node and delete the node that
the replacement value used to occupy. This results in invalidating an
iterator that is pointing to a node different from the one that is
actually being removed.

I did not end up swapping nodes in the RemoveNode function because
the code encountering this exact problem had some other smells that put
the issue to rest when handled. Additionally, swapping the nodes is a
lot more work than just replacing the value in the node to be removed.

Normally unused code would be discouraged, but this is probably the
third time I have felt there is a need for this function. Now that there
is a working version of it, it would be nonsensical to not keep it
around for the future.