Everything within editor/ has been refactored to use the new Interface structure.
Before this commit, there was a file name Primary.cc that handled all of the
logic for multiple different editor windows (Core, Overview, Inspector). This
was all in single file because all of these windows were closely intertwined
with each other. For example, both the Core and Overview windows want to know
the id of the selected Space. This change fixes that by making each of these
windows an Interface that can have sub-Interfaces. If an Interface needs
information from one of its sub-Interfaces, the sub-Interface can be searched
for to acquire it. Most Interfaces are WindowInterfaces because they display
editor windows.

This Interface change got the editor code off the fast track to becoming
complete spaghetti. Some editor windwow code had to handle closing windows that
were opened from completely different editor windows. The ability of Interfaces
to have sub-Interfaces addresses that problem. If an Interface is closed, all of
its sub-Interfaces are too. This means an Interface never needs to worry about
closing its sub-sub-...-Interfaces.

Another major advantage of this refactor (that wasn't even intentional at first)
is its effect on the system for managing gizmos. The system was obsolete and
removed. In its place, there are now HookInterfaces. HookInterfaces are an
Interface flavor just like WindowInterfaces. In the same way that
WindowInterfaces are used for dispalying windows, HookInterfaces are used for
displaying and editing components. They are called Hooks because they allow
components to hook onto the editor. Because these Hooks are only instantiated
when inspecting a component, all of a component's gizmo code can instead be put
within that component's Hook. This competely removes the need for any system
specifically for gizmos. The other neat advantage is that HookInterfaces are
closed and cleaned up if any parent Interfaces are closed, just like other
Interfaces.

This refactor also got its fingers into component specific code. Comp::TypeData
no longer contains any function pointers for component editor and gizmo Hooks
because of the introduction of HookInterface. Additionally, the component
Registrar is now a root level file. This has been moved to the root of src/ for
a few reasons. One, it's an extremely important file because it makes using cpp
components possible. Additionally, the Registrar interacts with multiple
systems. It registers a type as a component and it registers that component's
editor Hook. This makes the top of src/ a good spot for it. The file is also
purposefully small so that when a lot of components begin to enter the mix, the
Registrar won't be littered with other code that could be cumbersome to look at.

This commit also contains some more minor changes that were important to address
while refactoring.

Varkor now prints out memory leak information after terminating when running in
debug mode.

ImGui::DestroyContent() was not being called and this resulted in many memory
leaks from ImGui.

Code used for processing components.vlk was moved to Comp/Type.cc because that
code is entirely about checking the integrity of the TypeData.

Editor::Camera does not move if LeftControl is held down because that key acts a
modifier key.