Vlk::Tokenizer now gives line information on errors and Vlk::Token no longer has a length field.
The reason for giving line information should be obvious. It's like
searching for a needle in a haystack without it.

Vlk::Token no longer has a length field because the length field was
never necessary. The length can instead be found using the character
pointer on the next token. This also means that whitespace tokens are
now included in the token stream because they will contain character
pointers important for finding lengths. Vlk::Parser will have to handle
ignoring them on its own. This was intended though. The Parser will now
also have the ability to print line information.

There are also some small cleanup changes in here.