Most instances of int have been convertered to size_t where applicable and ComponentAddress is now ComponentDescriptor.
Only recently was I enlightened as to why size_t should be used over int. This
commit most likely does not handle every place in the code base where this
change should occur, but it should hanlde a majority. Others will be cleaned up
as they are encountered. Even in cases where size_t is not really necessary, the
clarity it brings in communicating what certain code does is still a large
benifit.

Using the name ComponentAddress resulted in a lot of confusing code due to how
general the word address is. This has been a point of friction since
World::Space was written and only while handling the int to size_t conversions
did a better name come to mind. Descriptor is used in reference to file
descriptors. It's not exactly the same, but similar to how file descriptors
point to an entry in the file table, component descriptors point to an entry in
one of multiple component tables.

As usual, this also contains some general cleanup.