Value has replaced Pair as the core interface used to interact with Valkor.
The one and only reason Pair was the used as Valkor's main interface was
to give the user the ability to have the root Value be a Pair. In other
words, the first Value within a vlk file could have a key string. Setting this
first key was only possible through Pair's constructor.

Initially this feature was not at all difficult to add. I thought, "Why
not let the user have a key for the first Value in a vlk file so
they could bake an extra string into the key?" The decision didn't seem
to have any cost. It did have a cost though and that cost was that Pair
had to be the primary interface used for Valkor to allow the user to
fill out that first key if they wanted to. This feature really wasn't that useful
either because one could just add a Pair to the root ValueArray and have
its Value be whatever the key string would have been.

This change fixes all of that. Value is now Valkor's primary interface
and Pair is nothing more than just a fly on the wall that a user will
never need to interact with.