The Value token type has been renamed to TrueValue. Value's String type has also been renamed to TrueValue.
These changes are being made for clarity reasons. The name of the Value
token changed because Value can mean a few different things. By saying
TrueValue, it now refers to one specific thing.

The name of the Value's String type changed because the purpose of that
type is not limited to strings. It also stores numbers even though they
are in a string format. The aim with renaming the type to TrueValue is to
encapsulate all these different cases.