A Sprite component has been added for displaying images. Image assets have also been added.
The sprite component is very similar to the model component and the process of
rendering either of these component types is also very similar. Sprites will
apply a special scaling to the model matrix to account for the image's aspect
ratio.

There is currently another struct called Gfx::Texture that was being used for
image loading and storage, but it contains members used specifically for
Gfx::Models. This data should be moved to Gfx::Model or Gfx::Mesh in the future.
Gfx::Image was created to be a reduced version of Gfx::Texture and places where
Gfx::Texture is used should eventually be replaced by Gfx::Image so Gfx::Texture
can be deleted.

This also adds a default image for when the ImageId of a Sprite component does
not represent a valid or live Image. It's just a picture of a questionmark.