The AssetLibrary now takes care of default Assets.
If a user makes a request for an Asset with an invalid AssetId, the
Asset returned will be the default Asset for the Asset type. Code to
handle this was in multiple places, but mostly the Renderer. Now the
AssetLibrary handles the problem all on its own.

This will be used more extensively in the future. When an Asset fails to
load, the default Asset will be used in place of that Asset. This will
happen once Asset loading is on a different thread.