All TrueValues within a vlk file are now wrapped with single quotes.
The main drive for doing this is to handle strings that aren't
necessarily used as strings. Consider enums. I want something more than
just a number when enum values are saved to file. I'd prefer to have the
identifier for the enumation value. This update to Valkor makes that
possible.

This makes things a bit simpler for the tokenizer. It will no longer
need to be able to recognizer numbers at all. Since all TrueValues are
wrapped in single quotes, they are very easy to identify.

This also adds an escape character for TrueValues. The reason is to make
the usage of a single quote possible within a TrueValue. The escape
character '\' can also be used twice to represent an occurence of it in
the string, exactly as one would expect.