The AssetLibrary now separates required Assets from user Assets using different ranges of AssetIds.
Before this, whether an Asset was required or not was tracked on the
Asset itself using a bool. Even though this could provide a distinction
between required and not, it failed to make use of the potential
AssetIds that an Asset could be given.

Now the AssetIds for requried Assets are always negative, the
AssetIds for regular Assets are positive, and the default id is 0. This
is really just a pleasant way to distinguish between the different
categories without an extra boolean or confusion.

Additionally, there is no longer an invalid AssetId. The elimination of
this began in 126a54d, but this finishes the job.