Adds the VInit component function and a call to a component's default constructor before deserialization.
When a component contains something like std::string, it needs to be
constructed to construct the string. The code for deserializing
a space did not do this and it resulted in a crash when a std::string
component member was used. Now components are default constructed before
serialization.

If a component's default constructor sets the sensible defaults for the
component, there will be a double write during the deserialization of
the component. 1 write for the defaults and 1 write for the
deserialization. VInit was added for component initialization to get
around this double write.