This adds text rendering using Text components and Font assets.
stb_truetype is used for loading the font bitmaps. FreeType was going to
be used, but since the stb repo provides the font loading, it was used
instead.

The current implementation does not have good performance. A draw call is
performed for every character. All characters are uploaded to the gpu
even if they aren't used and god forbid other font sizes (pixel heights)
are used because the old font sizes will not be deleted until the asset
is purged. Importantly though, it's working and future commits will
improve upon the base provided here.