Text rendering only supports sdf text rendering.
Using traditional glyph bitmaps is an acceptable solution when
we know the exact resoultion that the glyph will be displayed with.
If they are displayed at higher resoltions, the pixels from the bitmap
will quickly start to become visible. Because I expect text to be
rendered at different distances and therefore take up more or less pixel
space, bitmaps did not provide a good degree of quality for excessively
large glyphs (one that takes up a large portion of the screen).

Using an sdf representation solves that problem resulting in high
detail glyph renders at nearly any scale. This removes the need to track
font sizes because only one sdf render is needed for each glyph no
matter how much space that glyph will take up on screen.

This also adds a color member for the Text component.