Default Assets are now put within the same map as all other assets of the same type.
In a future commit, Varkor will be able to skip initialization and cleanly exit
to handle command line argument errors. Since all of Varkor's initialization
will be skipped, glad's initialization will not happen, resulting in a segfault
whenever an OpenGL call is made. Since default Assets are kept on the stack,
their destructors get called no matter what and these destructors call OpenGL
functions to free Asset data from the gpu. When Varkor's initialization is
skipped, though, these destructors cause a segfault because of the OpenGL call.

This mends that by not storing the default Assets on the stack. They are instead
stored in the same maps that other Assets are stored in, preventing any
unnecessary destructor calls.