Commandline arguments are now supported and the first allows a user to instantly load and select a space.
This can be done directly through the bvarkor build script as well.

This change required that the CoreInterface constructor be made into an Init function. The constructor cannot be used because initialization of the Options from commandline might stop Varkor from running, therefore also skipping over Varkor's purge step. All of the sub-interfaces stemming from the CoreInterface are freed during that step. So, since some sub-interfaces were being created in the CoreInterface constructor, memory leaks would occur if initialization of Options resulted in Varkor ending early.

The open, close, and find interface functions are now public, making it possible to add and modify sub-interfaces as long as the parent interface is available. This is used to open an OverviewInterface on the CoreInterface if the load-editor-space option is used. I believe these functions were under protected because the early version of the current iteration of the editor was very picky about when sub-interfaces were opened or closed. This opening and closing happens at one defined moment now though, so it should no longer be a problem.

This also adds ValuedResult. It's exactly the same as Result, but a value can also be included so it can be returned from functions that might return an error.