Result and ValuedResult have been moved to their own file, Result.h. ValuedResult has also been changed to ValueResult because I like the name more.
Considering the current extensive use of the Result structure and likely
ValueResult in the near future, these structures are deserving of
their own file in the root of src/.

They are no longer part of the Util namespace too. Having them in that
namespace was annoying in the first place. Now that Results are used
very often, it's an idiom of Varkor and having it in the Util namespace
resulted in unnecessary specification when it was used.