The component VInit function is only called if the component type defines it.
The Table unit test was not passing. Since the sample components do not
define VInit, the call to VInit in Table's add function caused a crash.
Now a check is performed to make sure VInit is defined every time a
component is added. I wonder if this has performance implications, but I am
not really in a position to account for that because of lack of scale
and need of refactoring. More importantly, this change allows the Table
and Space unit tests to pass and makes VInit optional.