This reworks how models are loaded and stored using a method that's more modular.
Originally, all meshes had the same format for their vertex buffers. Each vertex had a position, normal, and texture coordinate. If the texture coordinates or normals weren't used, it didn't matter. The number of bytes in single vertex would still be enough for all attributes. The major thing this change introduces is a fix to that problem. If a mesh doesn't contain an attirbute, then the bytes that would be used for that attribute will not be present in a vertex of that mesh anymore.

Though multiple texture coordinates can be put into a single vertex, there is only one attribute for texutre coordinates. Expanding this to allow for more texture coordinates is outside the scope of this change, but this is something that should be considered in the future once a need for it arises. The lack of focus on textures is very evident in the implementation of Gfx::Material because of the few number of texture types and loose rules about how textures are ordered.

Model::DrawInfo also eliminates the need for the Model's draw function. This gets the rendering calls outside of Gfx::Model and Gfx::Mesh so they can be called alongside the rest of the rendering code. This will also make instancing more easy in the future since this technique is more open to sorting the data into buckets.

This also adds a move contructor and assignment operator for every single asset type even when not necessary. Asset types should never be copied since they are just references to data stored on the gpu. They should only be moved. If they are copied, the asset copied from will be destructed and also invalidate the references in the copied version.

This also updates all default models to objs. Fbx files have some strange scaling standards and Obj files are plaintext. None of objs contain anything but position and normal attributes.