Error now contains a string that acts as a log of all errors.
Expected runtime errors were previously displayed with a popup
interface. Besides being an annoying way to feed imporant information
to the user, this method will not work when multiple errors occur at
the same time. Presenting errors as a log also allows the user to
ignore them until they are ready to address the issues.

The largest issue is that errors cannot be copied from the log window.
This will be addressed in the future when imgui is replaced with a custom
solution.