Assets are now loaded on a separate thread when they are needed.
This resulted in some major changes to the initialization of Gfx::Model. Most data uploaded to the gpu with OpenGL will be shared across threads when thread contexts are shared. One exception to this are vaos. Vaos must be created on the thread that will perform the rendering calls. Because of this, Gfx::Model has a unique initialization process. Almost everything is done on the init thread, but every Gfx::Mesh on a model must undergo finialization (creating the vao) on the main thread. Only after all meshes have been finalized will the model be ready to use for rendering.

There was a bug where a single mesh wouldn't render even though other meshes on the same model do. This bug showed up inconsistently, making me pretty certain that it had something to do with the interaction between the main and init threads. The bug was addressed by first reserving all of the space needed for the Gfx::Model::mMeshes vector before filling it in. Because the main thread might access this vector while the init thread is modifying it, there was an opportunity for the init thread to delete and recreate the allocation while the main thread was using it. I believe this fixed the bug, but since it showed up inconsistently, only time will truly tell.

Gfx::Shader's initialization process also underwent a chage. Instead of searching for a uniform's location every time it's going to change, the location of all uniforms is found during shader initialization. This results in a few things. First, all of the supported uniforms for Varkor shaders are explicitly listed in the Gfx::Shader header and implementation. Whether they will be set or not is a matter of context. Additionally, Gfx::Shader no longer has Set* functions for uniforms. Instead, the location of a desired uniform is retrieved from main memory and it's set using the appropriate glUniform* function. The best part about this is that all gl calls made during rendering are located in the same place. I hope that this will make rendering optimizations easier in the future.

Gfx::Material was simplified by putting all images of the same texture type into a group. This is easier to follow than stuffing all of the images into a single vector.

The Error namespace now contains a mutex for the error log because it's possible for errors to come from multiple threads. This mutex is being used in any place that the error log is accessed or changed.