VarkorMain has been replaced with VarkorInit, VarkorRun, and VarkorPurge.
These functions are important because Varkor projects need to able to perform operations between these steps.

These functions already existed in the past, but were replaced with VarkorMain in commit 5ed1459a. That commit introduced commandline options. The option initialization could fail and return an unsuccessful result. Reducing things to a single function was done so we could quickly dodge all of the other engine code. This was unnecessary though, because VarkorInit was able to simply return that result and allow main to perform the early termination.