The build scripts have been shifted around and slightly generalized to make them more easily usable in sub-projects.
The most important change here is the inclusion of the "target" build specification environment variable. This will allow a sub-project to use working/build.bat without needing to change things from "varkor" to whatever the main target of their project is.

This also changed the readme to say that Varkor is a static library engine. The rapid build section of the readme was updated to reflect the already mentioned changes too.