Functions removed from the previous commits have been re-added to make the World::Table test functional.
It's very likely that World::Table will be moved to world/Space.cc in the future. This means that the World::Space unit test will cover World::Table. This really should be the case, because the functionality of Space and Table depend on each other. That's beyond the scope of my current goal of adding dependencies though, so this change just focuses on making the unit test for World::Table functional again.