Components can now have dependencies.
If a component has dependency components, those dependencies will be added when the component is added to a space member. Likewise, if a component with dependants is removed, all of the dependants will also be removed. I am unsure about whether this approach is better than having users explicitly add and remove all required component types within their code. Since I am unsure, I am choosing to go down the automatic route and if there is evidence that this method is not adequate for some reason, it will be reworked to require explicit adding a removal.

This also adds the Registration test to the World::Space unit test. This is to make sure that the TypeData for test components is accurate. Ideally, this should be tested by generating a components.vlk file within the test, but that would require the unit testing framework to handle multiple golden files and that's far beyond the scope of this change.