ValueArray Vlk::Values can now contain PairArray types as elements.
The fact that this didn't work beforehand was just the combination of a few bugs. Firstly, the ReachedThreshold function assumed it would only need to deal with ValueArray types and not PairArrays. The other issue existed in Vlk::Parser. The ParsePairArray function did not assume it could be called while the top of the ValueStack is a ValueArray. A few of the other Parser functions were changed to reflect the same format the ParsePairArray function now has.

This also updates the Vlk::Value unit test to make sure these changes work. The Vlk::Explorer unit test was also updated. This update is probably not really necessary, but a bit more testing shouldn't hurt.