components.vlk has been updated to use a ValueArray with PairArray elements for storing component data.
This was the reason for d892d6f. The components are being stored this way because it more accurately reflects how the live component data is stored. It is not stored in a map where type names act as keys for different entries. It's stored in an array and the TypeData's index in that array is the id of the type. The components.vlk file now reflects this. Even though it's not necessary, the ids for components are also stored in the file to make referencing what component has a certain id easier.

This also updates the GetTypeId function that takes a component name and returns its id. The LogAbort for getting the id of a non-existant name has been removed because this function should not crash Varkor. This is absolutely the case for component assessment and space loading, the only two places where the function is currently called.