The total run time of the engine is now tracked by Temporal.
There was no reason for the Framer to store this value because it was not using it in any meaningful way. Temporal doesn't use it in a meaningful way either, yet, but it likely will in the future. More importantly, the value is now under a namespace that makes sense.

Temporal was also given an Update function. This happens right after Frame::Start because anything that happens inside of Update is part of the frame and that time needs to be accounted for, especially if the function is expanded in the future.