Added Comp::AlphaColor for modifying the uAlphaColor shader uniform.
I am currently working a subproject that needs to be able to modify this uniform value. This solution isn't necessarily perfect because it requires a specific addition to the renderer. Users won't necessarily have the same power unless they fork the engine. It will eventually be important that users have a more natural way to modify uniform values per object, but now is not the time.