World::Purge has been added to clear all of the existing Spaces.
Not having this lead to an unnecessary abort after exiting main. The Comp::nTypeData vector was cleared before the World::nSpaces vector and the Space destructor expects nTypeData to be filled out. Since we don't have control over which vector's destructor is called first, we need to clear the nSpaces vector before exiting main to guarantee that the nTypeData vector has not yet been cleared.