World::nCentralUpdate and World::nSpaceUpdate have been reexposed for use in subprojects.
CentralUpdate is currently being used in a subproject for testing purposes. SpaceUpdate is not, but it could be useful for some reason. Whether these functions have long term purpose is very much unknown, but reexposing and using them does not involve any sort of refactor. Since they aren't a liability, there shouldn't be any issue with them existing and they can be removed if and when their drawbacks become more clear.