World now stores all of the existing Spaces in a Ds::List rather than a Ds::Vector.
There are quite a few reasons for this change. Spaces are large structures that store other structures referencing data that describes a Space. More generally, they are a collection of pointers. The World does not need to store Spaces in a vector because we are not concerned about cache coherency. We should expect cache misses with either vector or list because of the structure's nature.

More importantly, though, the order of Spaces within the World matters and a user should have control over where a Space is in the list of Spaces that make up the World. The Renderer, for example, renders all the Spaces from front to back. Likewise, the World updates all of the Spaces from front to back. By storing all of the Spaces in a list, we allow the user to add and remove Spaces from the list at any desired location, giving them control over render layer and update order.

Not only that, but adding and removing at any location requires no movement of other Spaces within the World. This means that once a Space is created, that Space will always have the same address until it is removed. This makes World::Object finally useful since it can rely on Spaces with consistent memory addresses. This is why Object has spread to many files in this commit.

Those are the main things to be aware of, but here are some other small details that are a consequences of this change.

We need to const_cast World::Space within Renderer because World::Object uses a non-const pointer for the Object's Space. This is acceptable because the functions don't modify the Space despite casting away the const.

The structures that contain Space pointers as members are tied to the lifetime of the Space they contain.