Comp::Camera has been added.
This component shifts the responsibility of projection matrix creation from Viewport to a component that the user has control over. The intention is to give users control over what the projection matrix is, however, only perspective projections are currently supported.

An important thing to note is that Comp::Camera is dependant on Comp::Transform but the camera's corresponding Transform component is not accesssed anywhere within the Camera component. This is because we cannot guarantee that the Camera instance is within a space. Notice that Editor::Camera contains an instance of a Camera and a Transform. This is the primary reason why the Camera component cannot use a World::Object to access the Transform component. It must instead take in transformation values as parameters. This could be addressed by making an editor space to contain this object or by having Comp::Camera contain Comp::Transform. The latter, however, would raise many issues for the current state of editor hooks since the Transform component would be "hidden" within the Camera component.