Quaternion output now uses formal quaternion formatting rules.
This also creates math/Print.h and Print.cc. The purpose of these is to serve as a place to print all math structures. With this, Varkor code has easy access to math printing functions for debugging purposes. Quaternions are the first to be added, but others will follow when needed. This is much better than placing the quaternion printing function in its unit test.