Matrix4 now has a function for creating orthographic projections.
Orthographic projections are necessary for spaces that are meant to be viewed in 2D. With this projection type, elements at the edge of the window won't be unnecessarily skewed.

For some bizarre reason, the parameters of the perspective function in math/Matrix4.h were in the incorrect order. These and a small nitpick in the function itself have been updated.