Cameras now support the use of orthographic projections.
Some extra things have been done to ensure features working with perspective projections also work with orthographic projections. There is currently no need for it, but Camera::StandardToWorldPosition has been updated because of this. Whether the change worked or not was tested with the Transform gizmo. This is why the StandaredPositionToRay function was added. That change is part of this commit because it resulted in some simplification.