All spaces are now rendered to separate framebuffers before being rendered to the primary framebuffer.
This will allow for the easy application of post-processing effects in the near future, such as anti-aliasing and bloom.

This change also removed the space rendering operation from the overview interface. This means that all rendering operations, except for the gizmos, occur in the same place, making them easier to keep track of.

This also fixes a minor issue with object picking. It wouldn't work unless the inspector interface was open, now it does.