Gfx::Renderer now has a Purge function for clearing the space framebuffers.
This is necessary because a LogAbort would occur when closing a Varkor subproject. This happens because of Gfx::Framebuffer::smFullscreens. This is where all fullscreen framebuffers are tracked so they can be resized when a window resize occurs. However, since it's a static value, we have no way to guarantee when its constructed or destructed in relation to other namespace globals. A LogAbort occured because smFullscreens was cleared before Gfx::Renderer::nSpaceFramebuffers was cleared, which meant the space framebuffer pointers could not be found in anymore smFullscreens.