Component VInit functions now take the owner object as a parameter.
A component type in a sub-project needed access to its owner so it could access other components on the owning object. This change is a telling sign that the owner object should always be accessible from any of the component functions, but that's more difficult to achieve because it requires insight.