Varkor projects no longer require files to be copied into their directory.
Instead of a Varkor project having its own working directory, the project's executable is placed in Varkor's working directory. This way the executable sits alongside any required dlls and the vres/ directory.

Though copying files is not required, this did require the addition of a project directory option because the exectable needs to know where it can find all of the resource files that are specific to a project.