The Transform gizmo has been separated into 3 different gizmos.
This means we don't need a transform component in order to use the transformation gizmos. This is extremely useful because we can now use a gizmo for only a translating, scaling, or rotating. This wasn't possible earlier because the 3 gizmos were put into a single package, which made modifying a single part of a complete transformation impossible.

Gizmos are now responsbile for suppressing object picking in the main editor manually. Handling this with a return type was a bit ridiculous and it's now completely explicit when object picking is disabled.

This also adds an editor space. The editor space is where all editor elements are placed so they can be rendered together in a single call. This is where all of the objects used to represent the gizmo handles now exist.