Asset initialization has been simplified.
The process of loading models was overly complex. Models were the only asset type that needed finalization. This lead to the asset library having an explicit implementation for loading models. This is no longer the case. Instead, meshes hold onto the data needed for finilaztion (it's only 4 bytes). This has lead to all asset types having a finalize function, and I think this is sound. We have one function that does all processing that can be accomplished on a separate thread (Init), and another for processing that needs to be done on the main thread (Finalize). If an asset doesn't need to do anything on the main thread, then its implementation for Finalize can be left empty.

This also gets rid of the stingent requirements needed to make a type usable as an asset resource. This does sacrifice the labels for what the paths in the array of file paths given to the Init function represent, but it's not much of a loss, because I haven't seen any benefit from it. Additionally, it shouldn't be necessary for assets to define so many values related to paths when it's not guaranteed that an asset type will even use file paths for initialization. This will be the case for the newest addition to the asset library in a future commit, meshes.

I'm just tired and rambling at this point. Good night.