Line model generator.
I've been building this in a sub-project for a while now. The code isn't necessarily the cleanest and there are ways to improve it. For example, normals could be added, the tips of the arrow heads could be at the start and end points passed into the function, and the number of sides could be arbitrary. This works for what I need now, though. Bracket time.