Images will now use wrapping when sampled and Model importing now longer flips uvs.
The uv flippage was for the backpack asset that has sat in res/ for ages. I'm moving on to testing with the sponza model instead.

The clamp to edge for the image initalization was somewhat arbitrary and the sponza model expects it to be repeated, so it has been switched. Tbh, changing these values is probably going to need to be possible on an asset by asset basis, but that's something I'll worry about when and if I get there.